% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_man.R
\name{conv_man}
\alias{conv_man}
\title{Reverse scale of selected items using the key provided. Change existing items.}
\usage{
conv_man(df, x, items, key)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a character object}

\item{items}{a character vector}

\item{key}{an integer object}
}
\value{
a dataframe
}
\description{
Reverse scale of selected items using the key provided. Change existing items.
}
\examples{
d <- data.frame(id=c("josette", "paul", "louis"), be8_1=c(1,2,5), be8_2=c(2,3,5),be8_3=c(1,5,8))
conv_man(d,"be8",c(1,2),8)
}
