% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{pairing}
\alias{pairing}
\title{Strict pairing of ids through time 1 and time 2}
\usage{
pairing(df, id, date)
}
\arguments{
\item{df}{a dataframe}

\item{id}{a column name}

\item{date}{a column name}
}
\value{
a dataframe
}
\description{
Strict pairing of ids through time 1 and time 2
}
