% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\title{Create plots for each score between time 1 and time 2}
\usage{
plotting(df, date, var)
}
\arguments{
\item{df}{a dataframe}

\item{date}{a character object}

\item{var}{a character vector}
}
\value{
a plot
}
\description{
Create plots for each score between time 1 and time 2
}
