% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_groups.R
\name{plotting_groups}
\alias{plotting_groups}
\title{Create plots for each score intrasubject (time) and intersubjects (condition)}
\usage{
plotting_groups(df, date, group, var)
}
\arguments{
\item{df}{a dataframe}

\item{date}{a character object}

\item{group}{a character object}

\item{var}{a character vector}
}
\value{
a plot
}
\description{
Create plots for each score intrasubject (time) and intersubjects (condition)
}
