% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoring}
\alias{scoring}
\title{Automatically create scores for many categories of items}
\usage{
scoring(df, x)
}
\arguments{
\item{df}{a dataframe}

\item{x}{a character vector}
}
\value{
a dataframe
}
\description{
Automatically create scores for many categories of items
}
